Big structural error which we should consider before moving on
Since we removed movieDataBase, we have to start forward declaring StoreDriver and once we do that, a sort of circular dependency occurs since StoreDriver also includes Transactions. There are 2 ways to solve this, move the required objects/functions into a new file (basically re-creating movieDataBase but slightly different) or somehow get rid of one or the other file (StoreDriver or Transactions) to stop including/forward declaring the other.